distinct player class
